+++ History.md
@@ -5,6 +5,7 @@ v2.4.0
   * appCache has been disabled on non-secure connections in modern browsers
 * Fix for startup directory creation #59
 * Fix for memory leak and improvements to file upload cancellation #61
+* Support upload of large files (< 1.5GB) #63
 
 v2.3.1
 ======